requirements-dev.txt
# Development dependencies for Life Dashboard

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
behave>=1.2.6
selenium>=4.11.0

# Code Quality
ruff>=0.1.6
mypy>=1.7.1
django-stubs>=4.2.0
import-linter>=1.12.0
pre-commit>=3.5.0

# Development Tools
pydeps>=1.12.0
django-extensions>=3.2.3

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Data Analysis (for future analytics features)
pandas>=2.0.3
numpy>=1.24.3
scikit-learn>=1.3.0

# API Development (for future API features)
djangorestframework>=3.14.0
drf-spectacular>=0.26.0

# Async Task Processing
celery>=5.3.0
redis>=4.6.0

# Environment Management
python-dotenv>=1.0.0

# Validation
pydantic>=2.4.0
pydantic[email]>=2.4.0
